♻️ rename utils in recordsPerFullSnapshot
