+++ packages/rum/test/recordsPerFullSnapshot.ts
